server/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MongoDB async driver
motor>=3.3.0

# Image handling
Pillow>=10.0.0

# Embedder (ResNet-50 backbone)
torch>=2.2.0
torchvision>=0.17.0

# HTTP client (Shopify Admin API)
httpx>=0.27.0

# Numerical ops
numpy>=1.26.0

# Env-file support
python-dotenv>=1.0.0

# Multipart form uploads
python-multipart>=0.0.9
